[Pramod/ThoughtWorks] Issue 2251 - using Short for groupId leads to overflow problem. Changed to Integer (https://mifos.dev.java.net/issues/show_bug.cgi?id=2251)
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@14119 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/src/org/mifos/application/accounts/loan/struts/action/LoanAccountAction.java
@@ -487,8 +487,7 @@ public class LoanAccountAction extends AccountAppAction {
 				setBusinessActivitiesIntoSession(request);
 
 				List<ClientBO> clients = clientBusinessService
-						.getActiveClientsUnderGroup(customer.getCustomerId()
-								.shortValue());
+						.getActiveClientsUnderGroup(customer.getCustomerId());
 				if (clients == null || clients.size() == 0) {
 					throw new ApplicationException(
 							GroupConstants.IMPOSSIBLE_TO_CREATE_GROUP_LOAN);
@@ -1251,8 +1250,7 @@ public class LoanAccountAction extends AccountAppAction {
 			List<LoanBO> individualLoans = loanBusinessService
 					.getAllChildrenForParentGlobalAccountNum(globalAccountNum);
 			List<ClientBO> activeClientsUnderGroup = clientBusinessService
-					.getActiveClientsUnderGroup(customer.getCustomerId()
-							.shortValue());
+					.getActiveClientsUnderGroup(customer.getCustomerId());
 			List<LoanAccountDetailsViewHelper> clientDetails = populateClientDetailsFromLoan(
 					activeClientsUnderGroup, individualLoans, businessActivities);
 			loanActionForm.setClientDetails(clientDetails);
